[Silver IV] Title: 숫자 카드 2, Time: 856 ms, Memory: 47828 KB -BaekjoonHub

+++ 백준/Silver/10816. 숫자 카드 2/숫자 카드 2.cc
@@ -5,16 +5,16 @@ using namespace std;
 int main() {
 	ios::sync_with_stdio(0), cin.tie(0), cout.tie(0);
 	int N, M;
-	map<long long, int> m;
+	map<int, int> m;
 	cin >> N;
 	while (N--) {
-		long long a;
+		int a;
 		cin >> a;
 		m[a]++;
 	}
 	cin >> M;
 	while (M--) {
-		long long a;
+		int a;
 		cin >> a;
 		cout << m[a] << " ";
 	}
